printf

+++ push_swap/libft/ft_atoi.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int	result;
 	int	sign;
@@ -32,6 +32,7 @@ int	ft_atoi(const char *str)
 	{
 		i++;
 	}
+	
 	while (str[i] >= '0' && str[i] <= '9')
 	{
 		result = result * 10 + (str[i] - '0');
